feat: added UserDAODatabase

+++ src/main/java/com/project/picpay/infra/reporitoy/UserRepositoryDataBase.java
@@ -9,7 +9,6 @@ import org.springframework.stereotype.Repository;
 
 import java.util.Optional;
 
-
 @Repository
 public class UserRepositoryDataBase implements IUserRepository {
 
@@ -21,7 +20,7 @@ public class UserRepositoryDataBase implements IUserRepository {
 
     @Override
     public void save(User user) {
-        UserModel userModel = new UserModel(user.getUser_id(), user.getName(), user.getDocument(), user.getEmail(), user.getPassword(), user.getType_user(), user.getAmount());
+        UserModel userModel = new UserModel(user.getUser_id(), user.getName(), user.getDocument(), user.getEmail(), user.getPassword(), user.getType_user(), user.getAmount().doubleValue());
         this.userJPA.save(userModel);
     }
